requirements.txt
requests
beautifulsoup4
selenium
psycopg2-binary
python-dateutil
pandas